modify action to cart_items_controller
