+++ app/controllers/carts_controller.rb
@@ -11,33 +11,4 @@ class CartsController < ApplicationController
     redirect_to :back
   end
 
-  def remove_item_from_cart
-
-    @cart_item = CartItem.find(params[:id])
-    @cart_item.destroy
-    redirect_to :back
-  end
-
-  def change_quantity
-    @cart_item = CartItem.find(params[:id])
-    addOrRM = params[:addOrRM]
-
-    if addOrRM=="+"
-      quantity = @cart_item.product.quantity
-      if quantity > @cart_item.quantity+1
-        @cart_item.quantity = @cart_item.quantity+1
-        @cart_item.save
-      else
-        flash[:alert] = "商品已经抢光了哦"
-      end
-    else
-      if @cart_item.quantity>1
-        @cart_item.quantity = @cart_item.quantity-1
-        @cart_item.save
-      end
-    end
-
-    redirect_to :back
-  end
-
 end
