+++ app/helpers/cart_items_helper.rb
@@ -1,0 +1,2 @@
+module CartItemsHelper
+end
