+++ test/controllers/cart_items_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class CartItemsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
